taxiApi/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
Django==5.0.6
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.52.1
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mysqlclient==2.2.4
mysqlconnect==1.0.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.1
